requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.6
cycler==0.10.0
Cython==0.29.24
Django==3.2.7
django-cors-headers==3.8.0
django-shortcuts==1.6
easyocr==1.4.1
idna==3.2
imageio==2.9.0
kiwisolver==1.3.2
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.2
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
pdf2image==1.16.0
Pillow==8.2.0
pycodestyle==2.7.0
pyparsing==2.4.7
pytesseract==0.3.8
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
sqlparse==0.4.2
tesseract==0.1.3
tesseract-ocr==0.0.1
tifffile==2021.8.30
toml==0.10.2
torch==1.9.1
torchvision==0.10.1
typing-extensions==3.10.0.2
urllib3==1.26.7
wget==3.2
